pyproject.toml
[project]
name = "summariser"
version = "0.1.0"
description = "Creates a daily report from 1000s of articles"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "fastapi>=0.128.0",
    "hdbscan>=0.8.41",
    "ipykernel>=7.1.0",
    "openai>=2.16.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.0.0",
    "tqdm>=4.67.2",
    "umap-learn>=0.5.11",
    "uvicorn>=0.40.0",
]

[tool.uv]
package = true

# Standard build system for src-layout projects
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
